put Appstyles in utils folder
