+++ screens/utils/AppStyles.js
@@ -1,0 +1,43 @@
+const AppStyles = {
+  color: {
+    main: "#5ea23a",
+    text: "black",
+    purple: "#8e44ad",
+    title: "#464646",
+    subtitle: "#545454",
+    categoryTitle: "#161616",
+    tint: "#ff5a66",
+    description: "#bbbbbb",
+    filterTitle: "#8a8a8a",
+    starRating: "#2bdf85",
+    location: "#a9a9a9",
+    white: "white",
+    facebook: "#4267b2",
+    grey: "grey",
+    greenBlue: "#00aea8",
+    placeholder: "#a0a0a0",
+    background: "#f2f2f2",
+    blue: "#3293fe",
+  },
+  fontSize: {
+    title: 30,
+    content: 20,
+    normal: 16,
+  },
+  buttonWidth: {
+    main: "70%",
+  },
+  textInputWidth: {
+    main: "80%",
+  },
+  fontName: {
+    main: "Noto Sans",
+    bold: "Noto Sans",
+  },
+  borderRadius: {
+    main: 25,
+    small: 5,
+  },
+};
+
+export default AppStyles;
