+++ screens/utils/index.js
@@ -1,7 +1,8 @@
 import Firebase from './Firebase'
 import { FirebaseProvider, withFirebaseHOC } from './FirebaseContext'
 import mutations from './mutations'
+import AppStyles from './AppStyles'
 
 export default Firebase
 
-export { FirebaseProvider, withFirebaseHOC, mutations  }
+export { FirebaseProvider, withFirebaseHOC, mutations, AppStyles  }
